Polishing code

+++ include/resources.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #define LUTSIZE 1024
 extern const uint16_t Sine1024_12bit [1024];
-extern const uint16_t PowerOf2_16bit [1000];
+extern const double PowerOf2_16bit [1000];
 extern const uint16_t Triangle1024_12bit [1024];
 extern const uint16_t Sine256_12bit [256];
 extern const uint16_t Sine512_12bit [512];
